+++ packages/runtime/src/index.ts
@@ -41,7 +41,7 @@ import { routerHistory as history } from './history.js';
 import KeepAliveOutlet from './KeepAliveOutlet.js';
 import ClientOnly from './ClientOnly.js';
 import useMounted from './useMounted.js';
-import { Suspense, useSuspenseData } from './Suspense.js';
+import { withSuspense, useSuspenseData } from './Suspense.js';
 
 export {
   getAppConfig,
@@ -82,7 +82,7 @@ export {
   ClientOnly,
   useMounted,
 
-  Suspense,
+  withSuspense,
   useSuspenseData,
 };